Modular version of the impulse model and first steps of the properties identification.
The impulse model have been updated to add the possibility to select which type of rail, sleeper, pad and ballast to use. The user can also enter the distance between the sleepers. Simplified clamps are also available. Finally, there is also a new module, development in progress, to automate the properties identification of the sleepers.