Modifications from feedback (debug & run simu via terminal)
The "debug mode" option was implemented for both simulation phases independently. If it's on, the terminals running Aster jobs stay open at the end of a process. The simulation resumes when they are closed manually by the user.

The script runSingleSimu.py was created to run a simulation using a terminal or a bash script based on a .json parameter file. Read script documentation for more details.

Some modifications were done in the GUI concerning sleepers and rails mesh/materials, but are not used yet. These 2 tabs can be ignored.