Directory structure modification
The structure of src/MultiSleeperModel directory was modified to make it clearer. All "development files" are located in DevFiles, and then either App, AsterFiles or Messages. So now the main directory of the model is much lighter. The GUI is lauched with MultiSleeperModel.sh, and a simulation can still be run in the terminal using RunSingleSimu.py.