Update of the impulse model - possibility to run without mesh check
The Impulse_command_line.py and the data.json files have been modified in order to give the possibility to the user to not open Salome for mesh verification.
If mesh verification is set to false, le calculation skip the check and all the steps are run one after the other without any breaks.