Rails, sleeper & sleeper spacing parameter implementation
The user can now choose any rail and sleeper meshes, and define their materials properties. For now, rails 60E1 and 54E2 are available, as well as sleepers B91IV (=for 54E2) and B91VI (=for 60E1). Now, rail meshes are also centered on the origin. All group names that are common with the Impulse model have been updated (see Doc, 3-sleeper cell mesh).

The user can also choose a USP mesh, and must indicate its thickness. Sleeper spacing was implemented as well. The default value is 600mm. At the beginning of both simmulation phases, the mesh of the rail is scaled along X to match the actual sleeper spacing.