Implementation of spatial variability
Task 1.2: Possibility to create other macroelements to model spatial variability of the parameters (materials properties).

Not implemented in the GUI yet. JSON files can be edited to add the keyword "macroEl2", whose value is a dictionnary containing the keys of the materials properties, which will differ from the nominal macroelement. In addition to materials properties, the key "slpShift" allows introducing a sleeper forward shift (in mm) to allow varying the spacing between sleepers. If this value is not zero, a symmetrical macroelement with (-1)*slpShift is needed and automatically created to be able to build the left side of the track. Finally, the dictionnary can contain the key "computeModeShapesPh1". During Phase 1, if there is a sleeper shift, meshes don't match and mode shapes have to be computed for each macroelement. If not, one can choose to compute the mode shapes of the 2nd macroelement, or to reuse those of the nominal macroelement (faster).

In addition to the key "macroEl2", the key "macroElSequence" must contain a list of integers indicating how macroelements are sequenced. Its length must be the number of sleepers nSlp. The integers can be 1 (nominal), 2 (the additional macroelement) or 3 (the automatically created symmetrical version of the 2nd macroelement).

See example in MultiSleeperModel/ test.json and see the list of editable properties in MultiSleeperModel/properties_macroEl2.txt

Other changes
 - The mesh is rebuilt in Phase 2. Before, the one from phase 1 was reused, so it was not possible to add groups of interest (for FRFs) after the execution of Phase 1
 - If the acoustic mesh is 2D, the total radiated power radiated through the surface is computed. If it is 1D: if the mesh has groups of single nodes, the acoustic pressure is computed at each of these nodes. Groups of several nodes are ignored. If the mesh has no groups, groups N1, N2, etc. are created and the pressure is computed at all nodes.
 - The pickle acoustic data output was useless so it was deleted. FRF output files provide enough detail.
 - Clamps are added to the meshes using a function to make the code much cleaner
 - The GUI handles if too much RAM is set (memory limit). (but not if too little...)